% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregrate_queries}
\alias{aggregrate_queries}
\title{Combine query files from PubMedScrapeR into a single list object}
\usage{
aggregrate_queries(files, normalize = TRUE, blacklist = c())
}
\arguments{
\item{files}{a vector containing query file names that are to be combined into a single object}

\item{normalize}{boolean to indicate if query term counts should be normalized by number of abstracts}

\item{blacklist}{a vector containing terms to exclude, typically very common words}
}
\value{
a list object where each element corresponds to the term count of a given query
}
\description{
Combine query files from PubMedScrapeR into a single list object
}
\examples{
library(JubRi)
query_dir <- file.path("example", "test_queries")
files <- list.files(query_dir, pattern = "*.txt", full.names = TRUE, recursive = TRUE)
queries <- aggregrate_queries(files = files, normalize = TRUE)
# lapply(queries, head)
}
