% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubri.R
\name{JubRi}
\alias{JubRi}
\title{JubRi: a method for literature-based keyword ontology}
\usage{
JubRi(x, db, to_test_critera = 0.25, background_size = 1000,
  p_adj_method = c("holm", "hochberg", "hommel", "bonferroni", "BH",
  "BY", "fdr", "none"), pval_threshold = 0.01, log2fc_threshold = 1,
  seed = 42, verbose = TRUE)
}
\arguments{
\item{x}{a character vector of genes IDs thought to be found in database}

\item{db}{a database of literature queries}

\item{to_test_critera}{quantile to cut the number of terms to test. 0 if wanting to test all.}

\item{background_size}{number of elements to randomly pull from the database to use as a background model}

\item{p_adj_method}{p-value correction method}

\item{pval_threshold}{p-value significance threshold}

\item{log2fc_threshold}{log2 fold change significance threshold}

\item{seed}{set the initial seed for background aggregration and p-value computation}

\item{verbose}{speak alot}
}
\value{
a list of data.frames, each element corresponding to the foreground, background and metadata data.frames
}
\description{
JubRi: a method for literature-based keyword ontology
}
