% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log}
\alias{log}
\title{Verbose logging function}
\usage{
log(msg)
}
\arguments{
\item{msg}{character log statement}
}
\value{
a character log statement tagged with current system time
}
\description{
Verbose logging function
}
