% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{metaplot}
\alias{metaplot}
\title{Create metaplot of results}
\usage{
metaplot(x, top_n = 30, to_label = c("PVAL_LOG2FC", "LOG2FC"),
  file = file.path(getwd(), "metaplot.png"))
}
\arguments{
\item{x}{output metatable from \code{JubRi()}}

\item{top_n}{number of top enriched terms to show on on metaplot panel}

\item{to_label}{significance levels to label terms}

\item{file}{output file}
}
\value{
metaplot
}
\description{
Create metaplot of results
}
