% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries}
\alias{queries}
\title{OUTPUT FROM PUBMEDSCRAPER}
\format{An object of class \code{"list"}}
\source{
SOURCE
}
\usage{
data(queries)
}
\description{
DESCRIPTION GOES HERE
}
\examples{
data(queries)
}
\references{
REFERENCES
}
\keyword{datasets}
