% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reshape_to_df}
\alias{reshape_to_df}
\title{Reshape a list of query entries to a dataframe}
\usage{
reshape_to_df(x)
}
\arguments{
\item{x}{a list of query entries}
}
\value{
an aggregrated data.frame with IDs as columns and TERMS as rows
}
\description{
Reshape a list of query entries to a dataframe
}
