% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wilcox}
\alias{wilcox}
\title{Computer P-value enrichment of query terms}
\usage{
wilcox(fg, bg, terms)
}
\arguments{
\item{fg}{foreground matrix}

\item{bg}{background matrix}

\item{terms}{a vector of character terms that are to be tested}
}
\value{
p-values
}
\description{
Computer P-value enrichment of query terms
}
